// src/app/interfaces/iweapon.ts
export interface IWeapon {
    id: number;
    model: string;
    brand_id: number;
    type_id: number;
    caliber: string;
    barrel_length: string;
    total_length: string;
    ammunition_storage_type: string;
    ammunition_storage_capacity: string;
    picatiny_rail: string;
    mlok_system: string;
    action: string;
    permalink: string;
    thumbnail: string;
    images: string[];
}